Fix build issues on Fedora 37
A number of problems were exposed when building on Fedora 37 using
RPMFusion packaged obs-studio. Some implicitly included headers needed
to be added. The touch control needed to use the older ->pos() method
for compatibility, and the byte |= operator needed to be modified.

Also update the documentation to cover building against Qt5.

Fixes: #138, Cannot build the plugin in Fedora 37

+++ src/ptz-visca.cpp
@@ -748,7 +748,7 @@ obs_properties_t *PTZVisca::get_obs_properties()
 void PTZVisca::send(PTZCmd cmd)
 {
 	pending_cmds.append(cmd);
-	ptz_debug("sending cmd, queuelength=%lli", pending_cmds.count());
+	ptz_debug("sending cmd, queuelength=%i", (int)pending_cmds.count());
 	send_pending();
 }
